Move template files from ui to core
