+++ js/core/templates/child_default_template.js
