+++ js/core/templates/function_template.js
@@ -1,5 +1,5 @@
 var TemplateBase = require("./ui.template_base"),
-    domUtils = require("../../core/utils/dom");
+    domUtils = require("../utils/dom");
 
 var FunctionTemplate = TemplateBase.inherit({
