+++ js/core/templates/template_engines.js
@@ -1,4 +1,4 @@
-import { extractTemplateMarkup } from "../../core/utils/dom";
+import { extractTemplateMarkup } from "../utils/dom";
 import { registerTemplateEngine } from './template_engine_registry';
 
 registerTemplateEngine("jquery-tmpl", {
